Move TurnPoint from wbp to survey
