+++ welleng/survey.py
@@ -19,7 +19,7 @@ from .utils import (
     get_xyz
 )
 from .error import ErrorModel, ERROR_MODELS
-from .exchange.wbp import TurnPoint
+# from .exchange.wbp import TurnPoint
 # from .exchange.csv import export_csv
 
 
@@ -623,6 +623,34 @@ class Survey:
             The path and filename for saving the text file.
         """
         export_csv(self, filename)
+
+
+class TurnPoint:
+    def __init__(
+        self,
+        md=None,
+        inc=None,
+        azi=None,
+        build_rate=None,
+        turn_rate=None,
+        dls=None,
+        toolface=None,
+        method=None,
+        target=None,
+        tie_on=False,
+        location=None
+    ):
+        self.md = md
+        self.inc = inc
+        self.azi = azi
+        self.build_rate = build_rate
+        self.turn_rate = turn_rate
+        self.dls = dls
+        self.toolface = toolface
+        self.method = method
+        self.target = target
+        self.tie_on = tie_on
+        self.location = location
 
 
 def interpolate_md(survey, md):
